requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
pennylane>=0.33.0
pennylane-lightning>=0.33.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
Pillow>=10.0.0
tqdm>=4.65.0
pandas>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
